refactor: reuse login form
